requirements.txt
python-dotenv
transformers
